CRM/requirements.txt
Django==2.2.19
django-bootstrap3==15.0.0
django-filter==2.4.0
django-tables2==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
importlib-metadata==2.1.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
zipp==3.5.0
